convert browser base test to node.js base test.
