+++ lib/rrule.js
@@ -13,22 +13,19 @@
  * https://github.com/jkbrzt/rrule/blob/master/LICENCE
  *
  */
-;(function (root) {
-  var serverSide = typeof module !== 'undefined' && module.exports
+/* global module, define */
 
-  var getnlp = function () {
-    if (!getnlp._nlp) {
-      if (serverSide) {
-        // Lazy, runtime import to avoid circular refs.
-        getnlp._nlp = require('./nlp')
-      } else if (!(getnlp._nlp = root._RRuleNLP)) {
-        throw new Error(
-          'You need to include rrule/nlp.js for fromText/toText to work.')
-      }
-    }
-    return getnlp._nlp
+;(function (root, factory) {
+  if (typeof module === 'object' && module.exports) {
+    module.exports = factory()
+  } else if (typeof define === 'function' && define.amd) {
+    define([], factory)
+  } else {
+    root.RRule = factory(root)
+    root.RRuleSet = root.RRule.RRuleSet
+    root.rrulestr = root.RRule.rrulestr
   }
-
+}(typeof window === 'object' ? window : this, function (root) {
   // =============================================================================
   // Date utilities
   // =============================================================================
@@ -205,7 +202,6 @@
         bits[6] || 0,
         bits[7] || 0))
     }
-
   }
 
   dateutil.Time = function (hour, minute, second, millisecond) {
@@ -216,6 +212,7 @@
   }
 
   dateutil.Time.prototype = {
+    constructor: dateutil.Time,
     getHours: function () {
       return this.hour
     },
@@ -395,6 +392,7 @@
   }
 
   Weekday.prototype = {
+    constructor: Weekday,
     // __call__ - Cannot call the object directly, do it through
     // e.g. RRule.TH.nth(-1) instead,
     nth: function (n) {
@@ -1630,6 +1628,7 @@
   }
 
   IterResult.prototype = {
+    constructor: IterResult,
     init: function (method, args) {
       this.method = method
       this.args = args
@@ -2261,19 +2260,19 @@
   RRule.RRule = RRule
   RRule.RRuleSet = RRuleSet
   RRule.rrulestr = rrulestr
+  return RRule
 
-  if (serverSide) module.exports = RRule
-  if (typeof ender === 'undefined') {
-    root['RRule'] = RRule
-    root['RRuleSet'] = RRuleSet
-    root['rrulestr'] = rrulestr
+  function getnlp () {
+    // Lazy, runtime import to avoid circular refs.
+    if (!getnlp._nlp) {
+      if (root && root._getRRuleNLP) {
+        getnlp._nlp = root._getRRuleNLP(RRule)
+      } else if (typeof require === 'function') {
+        getnlp._nlp = require('./nlp')(RRule)
+      } else {
+        throw new Error('You need to include rrule/nlp.js for fromText/toText to work.')
+      }
+    }
+    return getnlp._nlp
   }
-
-  if (typeof define === 'function' && define.amd) {
-    /*global define:false */
-    define('rrule', [], function () {
-      // Always return RRule for Backward Compatibility
-      return RRule
-    })
-  }
-}(this))
+}))
