--- tests/index.html
@@ -1,16 +1,0 @@
-<!doctype html>
-
-<title>rrule.js tests</title>
-
-<link href="vendor/qunit.css" rel="stylesheet" />
-<script src="vendor/jquery.js"></script>
-<script src="vendor/qunit.js"></script>
-
-<script src="../lib/rrule.js"></script>
-<script src="../lib/nlp.js"></script>
-
-<script src="utils.js"></script>
-<script src="tests.js"></script>
-
-<div id="qunit"></div>
-<div id="qunit-fixture"></div>
